Refactor chatbot
