+++ chatbot/chat/request.json
@@ -2,7 +2,7 @@
     "language": "ja-JP",
     "botId": "Chatting",
     "appId": "8667ad4d-c6b0-447c-92a8-486a41f6551a",
-    "voiceText": "たぶん制限的に多分でまだ制限あるよもうすぐ",
+    "voiceText": "のことだということだよ",
     "clientData": {
         "option": {
             "nickname": "たいちょう",
