+++ t2s/synthesize-input.json
@@ -5,7 +5,7 @@
         "speakingRate": "1.00"
     },
     "input": {
-        "text": "それは、いいですね"
+        "text": "たしかにコロッケはそうですね"
     },
     "voice": {
         "languageCode": "ja-JP",
